Add timeline view showing your pings and those of people you follow
This ended up being surprisingly little code, but it's dense code
which does a lot in very little space. Also, the subquery expression
used to make the timeline efficient is a feature which only came out
in Django in the last few months. Hooray for it; without it, I would
have needed to write my own paginator.